WayForward: Added DS and Didj file IDs lists
The line #s for the file lists correspond to the internal IDs in both the SCN (scene) and LYR (layer / map screen) files, and can be useful for pairing up sets. GBA titles don't have file lists I can pull from, though.

And yeah, the ordering is different for most (but not all) regional variants (if a region isn't listed, assume both US and European versions).