WayForward: Leapster support and palette tweaks.
Dr. RNG informed me that the Leapster formats were pretty close to the format used for the rest of WayForward's catalogue after all, so I took the time to add support for them (and one pair of relevant file IDs). *Now* that should be everything.

I also added a toggle for adjusting palettes so that they conform to a 0-255 curve instead of a 0-248 one (or 0-240 in Leapster's case, which is the whole reason I added this option as they're a lot darker than they should be otherwise since they only have four colour bits instead of five).